# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Go4Ride API",
    "description": "\nRide-hailing **rider app** API (Phase 0\u20132). Driver and admin APIs are out of scope for this service.\n\n## Authentication\n\n1. `POST /api/v1/auth/request-otp` \u2014 send OTP to phone (`debug_otp` returned when `OTP_DEBUG=true`)\n2. `POST /api/v1/auth/verify-otp` \u2014 exchange OTP for JWT pair; creates account on first sign-in\n3. Click **Authorize** and paste the `access_token` (Bearer, no prefix)\n4. `POST /api/v1/auth/refresh` \u2014 new token pair using `refresh_token` in body (no header)\n\n## Typical booking flow\n\n1. Reverse geocode (optional) \u2192 list ride types \u2192 estimate fare\n2. `POST /api/v1/rides` with optional `Idempotency-Key` header\n3. `WS /api/v1/ws/rides/{ride_id}?token={access_token}` for live updates\n4. Cancel, history, repeat, or fetch invoice as needed\n\n## WebSocket\n\n```\nWS /api/v1/ws/rides/{ride_id}?token=<access_token>\n```\n\nEvents are JSON payloads published on ride status changes (mock driver advances stages in dev).\n\n## Errors\n\nStructured JSON errors: `{\"detail\": \"...\", \"code\": \"...\"}` with HTTP 4xx/5xx.\n",
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "https://go4ride-api.onrender.com",
      "description": "Render (example production)"
    }
  ],
  "paths": {
...
    "/api/v1/auth/request-otp": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Request Otp",
        "description": "Send an OTP to the given phone. Works for both new and returning riders.\n\nThe client should follow up with `POST /auth/verify-otp`. `is_new_user`\ntells the UI whether to show an onboarding step (e.g. ask for name) after\nverification.",
        "operationId": "request_otp_api_v1_auth_request_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOTPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OTPSentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/verify-otp": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify Otp",
        "description": "Verify OTP and return JWTs. Creates the rider account if first-time.",
        "operationId": "verify_otp_api_v1_auth_verify_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOTPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh",
        "description": "Exchange a refresh token for a new JWT pair.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "description": "Revoke the refresh token (client should discard both tokens).",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "description": "Return the authenticated user's id, phone, name, and role.",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/addresses": {
      "get": {
        "tags": [
          "addresses"
        ],
        "summary": "List Addresses",
        "description": "List saved addresses; pass `lat`/`lng` to sort by distance.",
        "operationId": "list_addresses_api_v1_addresses_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string",
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lat"
            }
          },
          {
            "name": "lng",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string",
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lng"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddressResponse"
                  },
                  "title": "Response List Addresses Api V1 Addresses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "addresses"
        ],
        "summary": "Create Address",
        "description": "Create a saved address (max 10 per user).",
        "operationId": "create_address_api_v1_addresses_post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/addresses/{address_id}": {
      "patch": {
        "tags": [
          "addresses"
        ],
        "summary": "Update Address",
        "description": "Update label, coordinates, or default flag on a saved address.",
        "operationId": "update_address_api_v1_addresses__address_id__patch",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "address_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Address Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "addresses"
        ],
        "summary": "Delete Address",
        "description": "Delete a saved address.",
        "operationId": "delete_address_api_v1_addresses__address_id__delete",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "address_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Address Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/email/send-verification": {
      "post": {
        "tags": [
          "email"
        ],
        "summary": "Send Verification",
        "description": "Send a verification code to the rider's email (`debug_code` when in dev).",
        "operationId": "send_verification_api_v1_email_send_verification_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailVerificationSentResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/email/verify": {
      "post": {
        "tags": [
          "email"
        ],
        "summary": "Verify Email",
        "description": "Verify email and grant one-time wallet credit bonus.",
        "operationId": "verify_email_api_v1_email_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddressCreateRequest": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Label"
          },
          "address_line": {
            "type": "string",
            "minLength": 1,
            "title": "Address Line"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Lng"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "label",
          "address_line",
          "lat",
          "lng"
        ],
        "title": "AddressCreateRequest"
      },
      "AddressResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "address_line": {
            "type": "string",
            "title": "Address Line"
          },
          "lat": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Lat"
          },
          "lng": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Lng"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "address_line",
          "lat",
          "lng",
          "is_default",
          "created_at",
          "updated_at"
        ],
        "title": "AddressResponse"
      },
      "AddressUpdateRequest": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "address_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          },
          "is_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default"
          }
        },
        "type": "object",
        "title": "AddressUpdateRequest"
      },
...
      "EmailVerificationSentResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "debug_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debug Code"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "EmailVerificationSentResponse"
      },
      "EmailVerifyRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 4,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "EmailVerifyRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LogoutRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "LogoutRequest"
      },
      "OTPSentResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "expires_in_minutes": {
            "type": "integer",
            "title": "Expires In Minutes"
          },
          "is_new_user": {
            "type": "boolean",
            "title": "Is New User"
          },
          "debug_otp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debug Otp"
          }
        },
        "type": "object",
        "required": [
          "message",
          "expires_in_minutes",
          "is_new_user"
        ],
        "title": "OTPSentResponse"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
...
      "RequestOTPRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 15,
            "minLength": 10,
            "title": "Phone",
            "examples": [
              "9876543210"
            ]
          }
        },
        "type": "object",
        "required": [
          "phone"
        ],
        "title": "RequestOTPRequest",
        "description": "Single-step OTP request \u2014 used for both first-time sign-up and returning login.",
        "examples": [
          {
            "phone": "9876543210"
          }
        ]
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "is_new_user": {
            "type": "boolean",
            "title": "Is New User",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "user_id",
          "role"
        ],
        "title": "TokenResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyOTPRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 15,
            "minLength": 10,
            "title": "Phone"
          },
          "code": {
            "type": "string",
            "maxLength": 8,
            "minLength": 4,
            "title": "Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "fcm_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fcm Token"
          },
          "platform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform"
          },
          "referral_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 16
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Code"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "code"
        ],
        "title": "VerifyOTPRequest",
        "examples": [
          {
            "code": "123456",
            "name": "Ravi Kumar",
            "phone": "9876543210"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT access token from `POST /api/v1/auth/verify-otp`. Paste the token only (Swagger adds the Bearer prefix)."
      }
    }
  },
...
}